# Viewport collision detection