--- a/src/data/feature_index.js
+++ b/src/data/feature_index.js
@@ -11,9 +11,9 @@ const vt = require('@mapbox/vector-tile');
 const Protobuf = require('pbf');
 const GeoJSONFeature = require('../util/vectortile_to_geojson');
 const arraysIntersect = require('../util/util').arraysIntersect;
+const TileCoord = require('../source/tile_coord');
 
-import type CollisionTile from '../symbol/collision_tile';
-import type TileCoord from '../source/tile_coord';
+import type CollisionIndex from '../symbol/collision_index';
 import type StyleLayer from '../style/style_layer';
 import type {SerializedStructArray} from '../util/struct_array';
 import type {FeatureFilter} from '../style-spec/feature_filter';
@@ -38,7 +38,9 @@ type QueryParameters = {
     params: {
         filter: FilterSpecification,
         layers: Array<string>,
-    }
+    },
+    tileSourceMaxZoom: number,
+    collisionBoxArray: any
 }
 
 export type SerializedFeatureIndex = {
@@ -61,22 +63,22 @@ class FeatureIndex {
     rawTileData: ArrayBuffer;
     bucketLayerIDs: Array<Array<string>>;
 
-    collisionTile: CollisionTile;
     vtLayers: {[string]: VectorTileLayer};
     sourceLayerCoder: DictionaryCoder;
 
+    collisionIndex: CollisionIndex;
+
     static deserialize(serialized: SerializedFeatureIndex,
-                       rawTileData: ArrayBuffer,
-                       collisionTile: CollisionTile) {
+                       rawTileData: ArrayBuffer) {
+        const coord = serialized.coord;
         const self = new FeatureIndex(
-            serialized.coord,
+            new TileCoord(coord.z, coord.x, coord.y, coord.w),
             serialized.overscaling,
             new Grid(serialized.grid),
             new FeatureIndexArray(serialized.featureIndexArray));
 
         self.rawTileData = rawTileData;
         self.bucketLayerIDs = serialized.bucketLayerIDs;
-        self.setCollisionTile(collisionTile);
 
         return self;
     }
@@ -114,8 +116,8 @@ class FeatureIndex {
         }
     }
 
-    setCollisionTile(collisionTile: CollisionTile) {
-        this.collisionTile = collisionTile;
+    setCollisionIndex(collisionIndex: CollisionIndex) {
+        this.collisionIndex = collisionIndex;
     }
 
     serialize(transferables?: Array<Transferable>): SerializedFeatureIndex {
@@ -167,9 +169,11 @@ class FeatureIndex {
         matching.sort(topDownFeatureComparator);
         this.filterMatching(result, matching, this.featureIndexArray, queryGeometry, filter, params.layers, styleLayers, args.bearing, pixelsToTileUnits);
 
-        const matchingSymbols = this.collisionTile.queryRenderedSymbols(queryGeometry, args.scale);
+        const matchingSymbols = this.collisionIndex ?
+            this.collisionIndex.queryRenderedSymbols(queryGeometry, args.scale, this.coord, args.tileSourceMaxZoom, pixelsToTileUnits, args.collisionBoxArray) :
+            [];
         matchingSymbols.sort();
-        this.filterMatching(result, matchingSymbols, this.collisionTile.collisionBoxArray, queryGeometry, filter, params.layers, styleLayers, args.bearing, pixelsToTileUnits);
+        this.filterMatching(result, matchingSymbols, args.collisionBoxArray, queryGeometry, filter, params.layers, styleLayers, args.bearing, pixelsToTileUnits);
 
         return result;
     }