--- a/src/render/painter.js
+++ b/src/render/painter.js
@@ -2,7 +2,6 @@
 
 const browser = require('../util/browser');
 const mat4 = require('@mapbox/gl-matrix').mat4;
-const FrameHistory = require('./frame_history');
 const SourceCache = require('../source/source_cache');
 const EXTENT = require('../data/extent');
 const pixelsToTileUnits = require('../source/pixels_to_tile_units');
@@ -12,6 +11,7 @@ const VertexArrayObject = require('./vertex_array_object');
 const RasterBoundsArray = require('../data/raster_bounds_array');
 const PosArray = require('../data/pos_array');
 const {ProgramConfiguration} = require('../data/program_configuration');
+const CrossTileSymbolIndex = require('../symbol/cross_tile_symbol_index');
 const shaders = require('../shaders');
 const Program = require('./program');
 const RenderTexture = require('./render_texture');
@@ -45,7 +45,8 @@ type PainterOptions = {
     showOverdrawInspector: boolean,
     showTileBoundaries: boolean,
     rotating: boolean,
-    zooming: boolean
+    zooming: boolean,
+    collisionFadeDuration: number
 }
 
 /**
@@ -58,7 +59,6 @@ class Painter {
     gl: WebGLRenderingContext;
     transform: Transform;
     _tileTextures: { [number]: Array<Texture> };
-    frameHistory: FrameHistory;
     numSublayers: number;
     depthEpsilon: number;
     lineWidthRange: [number, number];
@@ -94,13 +94,12 @@ class Painter {
     _showOverdrawInspector: boolean;
     cache: { [string]: Program };
     currentProgram: Program;
+    crossTileSymbolIndex: CrossTileSymbolIndex;
 
     constructor(gl: WebGLRenderingContext, transform: Transform) {
         this.gl = gl;
         this.transform = transform;
         this._tileTextures = {};
-
-        this.frameHistory = new FrameHistory();
 
         this.setup();
 
@@ -113,6 +112,8 @@ class Painter {
 
         this.basicFillProgramConfiguration = ProgramConfiguration.createBasicFill();
         this.emptyProgramConfiguration = new ProgramConfiguration();
+
+        this.crossTileSymbolIndex = new CrossTileSymbolIndex();
     }
 
     /*
@@ -276,9 +277,7 @@ class Painter {
         this.imageManager = style.imageManager;
         this.glyphManager = style.glyphManager;
 
-        this.frameHistory.record(Date.now(), this.transform.zoom, style.getTransition().duration);
-
-        for (const id in this.style.sourceCaches) {
+        for (const id in style.sourceCaches) {
             const sourceCache = this.style.sourceCaches[id];
             if (sourceCache.used) {
                 sourceCache.prepare(this.gl);