--- a/src/render/vertex_array_object.js
+++ b/src/render/vertex_array_object.js
@@ -13,6 +13,7 @@ class VertexArrayObject {
     boundIndexBuffer: ?IndexBuffer;
     boundVertexOffset: ?number;
     boundDynamicVertexBuffer: ?VertexBuffer;
+    boundDynamicVertexBuffer2: ?VertexBuffer;
     vao: any;
     gl: WebGLRenderingContext;
 
@@ -32,7 +33,8 @@ class VertexArrayObject {
          indexBuffer: ?IndexBuffer,
          vertexBuffer2: ?VertexBuffer,
          vertexOffset: ?number,
-         dynamicVertexBuffer: ?VertexBuffer) {
+         dynamicVertexBuffer: ?VertexBuffer,
+         dynamicVertexBuffer2: ?VertexBuffer) {
 
         if (gl.extVertexArrayObject === undefined) {
             (gl: any).extVertexArrayObject = gl.getExtension("OES_vertex_array_object");
@@ -45,11 +47,12 @@ class VertexArrayObject {
             this.boundVertexBuffer2 !== vertexBuffer2 ||
             this.boundIndexBuffer !== indexBuffer ||
             this.boundVertexOffset !== vertexOffset ||
-            this.boundDynamicVertexBuffer !== dynamicVertexBuffer
+            this.boundDynamicVertexBuffer !== dynamicVertexBuffer ||
+            this.boundDynamicVertexBuffer2 !== dynamicVertexBuffer2
         );
 
         if (!gl.extVertexArrayObject || isFreshBindRequired) {
-            this.freshBind(gl, program, layoutVertexBuffer, indexBuffer, vertexBuffer2, vertexOffset, dynamicVertexBuffer);
+            this.freshBind(gl, program, layoutVertexBuffer, indexBuffer, vertexBuffer2, vertexOffset, dynamicVertexBuffer, dynamicVertexBuffer2);
             this.gl = gl;
         } else {
             (gl: any).extVertexArrayObject.bindVertexArrayOES(this.vao);
@@ -57,6 +60,14 @@ class VertexArrayObject {
             if (dynamicVertexBuffer) {
                 // The buffer may have been updated. Rebind to upload data.
                 dynamicVertexBuffer.bind();
+            }
+
+            if (indexBuffer && indexBuffer.dynamicDraw) {
+                indexBuffer.bind();
+            }
+
+            if (dynamicVertexBuffer2) {
+                dynamicVertexBuffer2.bind();
             }
         }
     }
@@ -67,7 +78,8 @@ class VertexArrayObject {
               indexBuffer: ?IndexBuffer,
               vertexBuffer2: ?VertexBuffer,
               vertexOffset: ?number,
-              dynamicVertexBuffer: ?VertexBuffer) {
+              dynamicVertexBuffer: ?VertexBuffer,
+              dynamicVertexBuffer2: ?VertexBuffer) {
         let numPrevAttributes;
         const numNextAttributes = program.numAttributes;
 
@@ -84,6 +96,7 @@ class VertexArrayObject {
             this.boundIndexBuffer = indexBuffer;
             this.boundVertexOffset = vertexOffset;
             this.boundDynamicVertexBuffer = dynamicVertexBuffer;
+            this.boundDynamicVertexBuffer2 = dynamicVertexBuffer2;
 
         } else {
             numPrevAttributes = (gl: any).currentNumAttributes || 0;
@@ -105,6 +118,9 @@ class VertexArrayObject {
         if (dynamicVertexBuffer) {
             dynamicVertexBuffer.enableAttributes(gl, program);
         }
+        if (dynamicVertexBuffer2) {
+            dynamicVertexBuffer2.enableAttributes(gl, program);
+        }
 
         layoutVertexBuffer.bind();
         layoutVertexBuffer.setVertexAttribPointers(gl, program, vertexOffset);
@@ -118,6 +134,10 @@ class VertexArrayObject {
         }
         if (indexBuffer) {
             indexBuffer.bind();
+        }
+        if (dynamicVertexBuffer2) {
+            dynamicVertexBuffer2.bind();
+            dynamicVertexBuffer2.setVertexAttribPointers(gl, program, vertexOffset);
         }
 
         (gl: any).currentNumAttributes = numNextAttributes;