Switching from tiled label placement to global viewport label placement. https://github.com/mapbox/mapbox-gl-js/issues/4704 WIP.

This PR dramatically changes our approach to collision detection. Instead of trying to pre-calculate collisions across a range of zooms in the background, we now calculate collisions synchronously in the foreground. Fade animations are no longer tied to zoom level, but instead based on time elapsed since collision occurred. This approach helps solve many problems:

 - More reliable collision detection between tiles (buffering/avoid-edges are no longer necessary)
 - Collision detection now works on labels from different sources (e.g. a GeoJSON set of labels added on top of a base map)
 - Calculating collision boxes at (or near) render time is much more accurate for pitched maps, line labels, sizes based on zoom functions, etc.
 - Fade in/out animations are no longer limited to zoom operations (i.e. collisions related to pitch/pan/rotate will all be animated now)

The main structural changes are:

 - `CollisionTile` is now a global object that lives only in the foreground. Insertions/queries into the collision index are now done entirely in viewport coordinates instead of tile coordinates.
 - The "placement" step no longer happens in the background thread, and static buffers for symbols are created during the "prepare" step.
 - The new foreground "placement" step is primarily responsible for updating a dynamic array of opacities that are used to animate collision states. For each vertex, the opacity state is expressed as an "opacity at time of last update" and a "target opacity". The symbol shaders animate towards the "target opacity" based on an animation start time set for each bucket.
 - It's now necessary to place a hard time limit on how long an individual placement call can take, to avoid frame stutter -- so full placement can now be split across multiple frames.
 - Collision features for lines are now represented in the collision grid as an array of circles, instead of as an array of boxes. This makes their behavior more stable under pitch/rotation operations.
 - Because fading is now tied to "collision time" instead of to zoom level, we need a way to tell when two symbols in tiles of two different zoom levels are "the same symbol", so that we can avoid triggering a fade in/fade out when we cross between zoom levels. This is primarily handled by the new `CrossTileSymbolIndex`.

TODO:

- [x] Handle re-ordering for overlapping symbols (requires re-generating static buffers on rotation -- might be time to address issue #2706?)
- [x] Fix remaining test failures/update tests to reflect new behavior
  - [x] Render tests
    - [x] 10 remaining render test failures for symbol-avoid-edges/tile-clipping (current behavior is actually better)/overlap-sorting
  - [x] Query tests
  - [x] Unit tests
- [ ] Performance improvements:
  - [ ] ~~Don't require full placement when a new tile is added -- instead, start progressive placement for the tile as soon as it's added, but don't start rendering until the first full placement has completed.~~
  - [x] Allow progressive placement to pause at the bucket level. (Currently placement pauses at the layer level, where a single dense layer of street labels on a large viewport can take ~8ms on my MacBook Pro)
- [x] Implement anti-aliasing for collision circle debugging (need them to look good for Studio)
- [x] Extend collision detection to a buffer around the viewport (so that animations for changes around the edge can be mostly finished animating before they're actually visible)
- [x] Update comments, rationalize class names (e.g. `CollisionTile`), remove dead code
- [x] Rebase on master
- [x] Flow-ify changes
- [x] Handle wrapped tiles in `CrossTileSymbolIndex`
- [x] Figure out home for `grid_index_experimental` (adds support for circle geometry queries; removes support for cross-thread serialization): either update https://github.com/mapbox/grid-index, make a new library, or just continue to include the code directly.
- [ ] Port to `gl-native`
  - [ ] Figure out how to continue doing tile-based collision detection in `api-gl`

This PR will enable further changes that we're not tackling yet:

 - Collision detection for labels with rotation-alignment map.
 - `text-offset` applied to collision detection
 - Adding collision features with arbitrary geometry
 - Cross-fading for layout property changes
 - Multiple label placement (https://github.com/mapbox/mapbox-gl-js/issues/5038 / https://github.com/mapbox/mapbox-gl-js/tree/multiplace)
 - Label density reduction: #5086
 - Reduce data usage by creating vector tiles without collision buffers
 - Fade out symbols from removed tiles (requires holding on to tile or at least symbol buckets over period of fade animation)

## Latest Benchmark

<img width="920" alt="screenshot 2017-09-29 11 15 44" src="https://user-images.githubusercontent.com/375121/31029703-a1653da8-a507-11e7-8f51-d25af1ba4a20.png">

/cc @ansis @nickidlugash @mollymerp