@nickidlugash 2017-08-21
@ChrisLoer A few notes based on my visual review of this branch today:

- Overall I think this is looking awesome 💯  Super excited to see this improved behavior on all our styles!

- I'm noticing a lot of labels that are initially visible and then disappear while panning. It looks like that is what would be addressed in this TODO item:
  > Extend collision detection to a buffer around the viewport (so that animations for changes around the edge can be mostly finished animating before they're actually visible)

  How will we calculate how big this buffer needs to be? Would we approximate it based on the default collision fade duration, and a reasonable panning speed? Would we be able to make a buffer large enough to resolve this issue for the background in very pitched maps?

- `text-offset` for labels along lines isn't taken to account in collision. Is this going to be part of further work listed?:

  >  Collision detection for labels with rotation-alignment map.

- As I mentioned in chat, might be worth us thinking about whether `collisionFadeDuration` should be a map property (as currently implemented in this branch), or a global style property (or both)? I could see this being a useful property in both.

- Now that you've identified exactly what updates will and won't be part of this initial PR, I'll open a couple more tickets for specific future changes so we can start tracking those discussions separately.