@ChrisLoer 2017-08-22
> How will we calculate how big this buffer needs to be? Would we approximate it based on the default collision fade duration, and a reasonable panning speed? Would we be able to make a buffer large enough to resolve this issue for the background in very pitched maps?

These are all good questions, and I don't know the answer. We should experiment with different values to find the best stability/performance tradeoff. I can try to make the padding a configurable value just so we can experiment more quickly.

> text-offset for labels along lines isn't taken to account in collision

Actually, I think that's just a bug -- there are two failing test cases for that too. It's on my list to look into, I've just been mentally filing it under "fix tests".

> As I mentioned in chat, might be worth us thinking about whether collisionFadeDuration should be a map property (as currently implemented in this branch), or a global style property (or both)? I could see this being a useful property in both.

:+1: