@ChrisLoer 2017-08-22
OK, the query tests are now passing and the ten remaining render test failures are all caused by the known avoid-edges/tile-clipping/overlapping-symbol-sorting issues. If we removed support for symbol-avoid-edges and stopped doing tile clipping for overlapping symbols, we'd be down to just two failures, both in `icon-translate-anchor`, caused by changing the sort order of overlapped symbols on a rotated tile.

> > text-offset for labels along lines isn't taken to account in collision

> Actually, I think that's just a bug -- there are two failing test cases for that too. It's on my list to look into, I've just been mentally filing it under "fix tests".

I was too optimistic here. The cause of those test failures was a bug in handling collision boxes for single-glyph labels. Actually handling text-offset correctly in collisions for line labels isn't included in this PR, but it should be much more tractable after this PR goes in. I've added it to "further work".