@ChrisLoer 2017-08-22
Benchmark results overall look pretty good, but `frame-duration` has gone up. A modest increase in the average may be acceptable, but for the zoom 13 test it caused at least one frame to take longer than 16ms:

master b203a40 | viewport-collision 1cff6c7
--- | ---
7.6 ms, 0% > 16 ms at zoom 13 | 8.4 ms, 1% > 16 ms at zoom 13

benchmark | master b203a40 | viewport-collision 1cff6c7
--- | --- | ---
**map-load** | 168 ms  | 126 ms 
**style-load** | 70 ms  | 67 ms 
**buffer** | 1,002 ms  | 1,007 ms 
**fps** | 60 fps  | 60 fps 
**frame-duration** | 5.3 ms, 0% > 16ms  | 6 ms, 0% > 16ms 
**query-point** | 0.55 ms  | 0.59 ms 
**query-box** | 39.98 ms  | 39.50 ms 
**geojson-setdata-small** | 1 ms  | 2 ms 
**geojson-setdata-large** | 123 ms  | 109 ms