@ChrisLoer 2017-08-25
I have a local stash that replaces the "flattened" arrays of collision boxes and collision circles with arrays of javascript objects. It's much easier to read. However, it also seems to be slower. My test methodology was to rotate a pitched full screen map with lots of road labels while recording with the CPU profiler, and then look at the percent of render time spent on the relevant code (looking at ratios is more reproducible than looking at absolute render time):

Before:

 Run| redoPlacement | placeCollisionCircles (self)
--- |--- | ---
First | 18.2% | 8.5% (2.9%)
Second | 16.7% | 7.5% (2.4%)
Third | 16.7% | 7.4% (2.5%)

After:

Run | redoPlacement | placeCollisionCircles (self)
--- |--- | ---
First | 21.0% | 11.5% (3.5%)
Second | 18.6% | 8.5% (2.7%)
Third | 23.4% | 13.5% (4.1%)

I don't have a good explanation for why this makes such a noticeable difference, but it looks like in this particular example it pushes up placement time by at least a millisecond.