@ChrisLoer 2017-08-25
I did four more test runs, this time with a build that represented circles/boxes as javascript objects, but with only number data members (i.e. instead of containing a `Point` member, they just contained an `x` and a `y`). Results looked pretty similar -- still a noticeable win for the earlier "flat" approach.