@mourner 2017-08-25
@ChrisLoer I also noticed that working with flat number arrays generally gives a huge boost to performance compared to arrays of objects. V8 treats them differently. Earcut and Delaunator both take advantage of this.