@ChrisLoer 2017-08-25
D'oh. I found that at least part of the cause of the flattened array being faster was that it had a bug that caused it not to place some collision circles that it should have (of course the bug derived from the awkwardness of the code!).

After fixing the bug, the difference is looking narrower, but the "flattened" approach is still reliably faster...

Before:

 Run| redoPlacement | placeCollisionCircles (self)
--- |--- | ---
First | 22.1% | 11.8% (2.3%)
Second | 21.4% | 11.0% (2.0%)
Third | 20.7% | 11.3% (2.0%)

After:

Run | redoPlacement | placeCollisionCircles (self)
--- |--- | ---
First | 22.7% | 12.7% (4.0%)
Second | 23.9% | 13.5% (4.0%)
Third | 22.6% | 12.8% (2.5%)