@ChrisLoer 2017-08-25
@mourner That's good to know, thanks! @kkaefer had theorized that v8 would be able to be clever about recognizing that objects were "plain ol' data" and do the flattening internally, but I guess not, or at least not yet...