@ChrisLoer 2017-08-28
@nickidlugash I pushed a change that extends collision detection past the edge of the viewport by 100 pixels in all directions. Unfortunately, there's no getting around that doing collision over a broader area makes it significantly more expensive (the cost scales at least linearly with the number of symbols included in the detection area). I ended up at padding of 100px because the hit to collision times wasn't *that* bad, and it seemed to eliminate most of the notable cases of label instability at the edge of the viewport. Maybe you could evaluate if it's good enough? If we can get away with an even lower value, that'd be great, too (just change the `viewportPadding` constant in `collision_index` to try different values).