@ChrisLoer 2017-08-29
@nickidlugash I just pushed a change that draws the collision circles with anti-aliasing, so they should be "ready" for use in Studio. I could use your help figuring out the color strategy though. The current black and blue (with fading to indicate unused collision boxes/circles) works well for me, and I think it looks reasonable. We could go back to green and red, but those are really difficult for my kind to figure out. Overall, I think the decision is a lot easier because we used to have four colors for the various states, but now there are just two colors.