@ChrisLoer 2017-09-05
After rebasing on master and combining with the expressions PR (#4777), the benchmark results look problematic. 😞 

In both density plots, the flatter line is `viewport-collision`.

Both branches seem to have a longer tail than we'd like -- I'm not sure what we expect is causing that on the expressions side (I'm assuming the changes on master are because of #4777, that could be wrong). On the `viewport-collision` side, I know the tail can be long because it's sometimes necessary to synchronously perform a full placement. The tail also seemed to get fatter, and I'm not sure why. I haven't found any obvious regressions in performance otherwise (i.e. average placement time hasn't increased), but I'll poke into that more.

## Before expressions

benchmark | master c48a56b | viewport-collision 67e3ca6
--- | --- | ---
**map-load** | 114 ms  | 82 ms 
**style-load** | 55 ms  | 54 ms 
**buffer** | 1,006 ms  | 977 ms 
**tile_layout_dds** | 962 ms  | n/a 
**fps** | 59 fps  | 59 fps 
**frame-duration** | 5.2 ms, 0% > 16ms  | 5.7 ms, 0% > 16ms 
**query-point** | 0.55 ms  | 0.61 ms 
**query-box** | 41.54 ms  | 42.29 ms 
**geojson-setdata-small** | 1 ms  | 2 ms 
**geojson-setdata-large** | 106 ms  | 52 ms 


![image](https://user-images.githubusercontent.com/375121/29896548-516a21d8-8d92-11e7-9e0f-efab7df0c0c3.png)

## After expressions

benchmark | master 8a674b8 | viewport-collision 669c9c3
--- | --- | ---
**map-load** | 87 ms  | 83 ms 
**style-load** | 110 ms  | 108 ms 
**buffer** | 1,052 ms  | 1,083 ms 
**tile_layout_dds** | 1,493 ms  | 1,453 ms 
**fps** | 60 fps  | 60 fps 
**frame-duration** | 6 ms, 0% > 16ms  | 7.1 ms, 4% > 16ms 
**query-point** | 0.83 ms  | 1.03 ms 
**query-box** | 68.71 ms  | 74.89 ms 
**geojson-setdata-small** | 7 ms  | 9 ms 
**geojson-setdata-large** | 141 ms  | 104 ms 
**filter** | n/a  | n/a 

![image](https://user-images.githubusercontent.com/375121/30087042-248c78e4-9253-11e7-8f05-5945fefc3606.png)

/cc @anandthakker